# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DevHub personal workspace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reminders:
    get:
      operationId: listReminders
      tags: [reminders]
      summary: List all reminders
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
    post:
      operationId: createReminder
      tags: [reminders]
      summary: Create a reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{id}:
    get:
      operationId: getReminder
      tags: [reminders]
      summary: Get a reminder by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
        "404":
          description: Not found
    patch:
      operationId: updateReminder
      tags: [reminders]
      summary: Update a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
        "404":
          description: Not found
    delete:
      operationId: deleteReminder
      tags: [reminders]
      summary: Delete a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Meetings ─────────────────────────────────────────────────────────────
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── AI Prompts ────────────────────────────────────────────────────────────
...
  /snippets:
    get:
      operationId: listSnippets
      tags: [snippets]
      summary: List all code snippets
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Snippet"
    post:
      operationId: createSnippet
      tags: [snippets]
      summary: Create a code snippet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnippetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snippet"
  /snippets/{id}:
    get:
      operationId: getSnippet
      tags: [snippets]
      summary: Get a snippet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snippet"
        "404":
          description: Not found
    patch:
      operationId: updateSnippet
      tags: [snippets]
      summary: Update a snippet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnippetUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snippet"
        "404":
          description: Not found
    delete:
      operationId: deleteSnippet
      tags: [snippets]
      summary: Delete a snippet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Bookmarks ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Reminder:
      type: object
      required: [id, title, dueAt, done, createdAt, updatedAt]
      properties:
        id:          { type: integer }
        title:       { type: string }
        description: { type: ["string", "null"] }
        dueAt:       { type: string, description: "ISO-8601 datetime" }
        done:        { type: boolean }
        createdAt:   { type: string }
        updatedAt:   { type: string }
    ReminderInput:
      type: object
      required: [title, dueAt]
      properties:
        title:       { type: string, minLength: 1 }
        description: { type: string }
        dueAt:       { type: string }
        done:        { type: boolean }
    ReminderUpdate:
      type: object
      properties:
        title:       { type: string, minLength: 1 }
        description: { type: string }
        dueAt:       { type: string }
        done:        { type: boolean }

    # ── Meeting ──────────────────────────────────────────────────────────
...
    Task:
      type: object
      required: [id, title, status, priority, createdAt, updatedAt]
      properties:
        id:          { type: integer }
        title:       { type: string }
        description: { type: ["string", "null"] }
        status:      { type: string, enum: [todo, in_progress, done] }
        priority:    { type: string, enum: [low, medium, high] }
        dueAt:       { type: ["string", "null"] }
        tags:        { type: array, items: { type: string } }
        createdAt:   { type: string }
        updatedAt:   { type: string }
    TaskInput:
      type: object
      required: [title]
      properties:
        title:       { type: string, minLength: 1 }
        description: { type: string }
        status:      { type: string, enum: [todo, in_progress, done] }
        priority:    { type: string, enum: [low, medium, high] }
        dueAt:       { type: string }
        tags:        { type: array, items: { type: string } }
    TaskUpdate:
      type: object
      properties:
        title:       { type: string, minLength: 1 }
        description: { type: string }
        status:      { type: string, enum: [todo, in_progress, done] }
        priority:    { type: string, enum: [low, medium, high] }
        dueAt:       { type: string }
        tags:        { type: array, items: { type: string } }

    # ── Prompt ────────────────────────────────────────────────────────────
...
    Snippet:
      type: object
      required: [id, title, code, language, tags, createdAt, updatedAt]
      properties:
        id:          { type: integer }
        title:       { type: string }
        code:        { type: string }
        language:    { type: string, description: "e.g. typescript, python, bash" }
        description: { type: ["string", "null"] }
        tags:        { type: array, items: { type: string } }
        createdAt:   { type: string }
        updatedAt:   { type: string }
    SnippetInput:
      type: object
      required: [title, code, language]
      properties:
        title:       { type: string, minLength: 1 }
        code:        { type: string, minLength: 1 }
        language:    { type: string, minLength: 1 }
        description: { type: string }
        tags:        { type: array, items: { type: string } }
    SnippetUpdate:
      type: object
      properties:
        title:       { type: string, minLength: 1 }
        code:        { type: string, minLength: 1 }
        language:    { type: string, minLength: 1 }
        description: { type: string }
        tags:        { type: array, items: { type: string } }

    # ── Bookmark ──────────────────────────────────────────────────────────
...