# lib/api-zod/src/generated/types/reminder.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DevHub personal workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface Reminder {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** ISO-8601 datetime */
  dueAt: string;
  done: boolean;
  createdAt: string;
  updatedAt: string;
}