# lib/api-zod/src/generated/types/snippet.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DevHub personal workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface Snippet {
  id: number;
  title: string;
  code: string;
  /** e.g. typescript, python, bash */
  language: string;
  /** @nullable */
  description?: string | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}