# lib/api-zod/src/generated/types/snippetUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DevHub personal workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface SnippetUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  code?: string;
  /** @minLength 1 */
  language?: string;
  description?: string;
  tags?: string[];
}